Update LightMQTT.swift
Completely refactored InputStream reading in background thread, as the previous implementation was causing BAD_EXC_ACCESS because of access to shared state from asynchronous code. In the latest attempt the readStream(inputStream: InputStream) function keeps a reference to its InputStream and reads available bytes in a while loop until stream state is not .open any longer.

If the MQTT client is reconnected resulting in new InputStream and OutputStream, the readStream loop waits until the inputStream.read() function times out. This seems to take at most 60 seconds. It then exits the loop and returns from the closure. All state is internal to the readStream function which prevents random crashes seen previously.