Update LightMQTT.swift
Refactor socketConnect into a more self contained openStreams that returns an opened inputStream and outputStream and works less with global state.

Capture outputStream for delayedPing, so that pings are sent to that specific stream in the dispatched recursive ping calls, not the current outputStream from the instance. Check that outputStream is still open before pinging.

Also capture outputStream for readStream. Although it's not used yet, this makes it possible to send replies to server using the corresponding outputStream when needed.