Update LightMQTT.swift
Fix a pretty bad bug that caused variable MQTT header bytes being read as fixed header, potentially causing all sorts of issues like disconnections. Now, once message length is known, the remaining bytes are read from the input stream for all message types, not only publish messages. This was not visible in my usage of the MQTT client, so the bug remained there for quite long.

Make writing to output stream asynchronous, so that it doesn't block the thread it's called from, and add a completion block so the caller can send several messages in sequence.

Input stream is checked for available bytes in a while loop before reading, so the read does not block for 60 seconds until it closes and the read loop returns. Specific input and output streams can now be disconnected, instead of the current ones stored in instance variables. This make stream handling much clearer and allows disconnected and closed stream reading to exit faster.

Add usleep to potentially tight while loops to prevent using a lot of CPU in bad network conditions or with MQTT servers publishing a lot of data. Not sure if this is necessary, though.

Change DispatchQueue QoS levels to .default, to make sure they don't block the main thread, and app UI stays responsive even with bigger message loads.

Add options for socketSecurityLevelKey and networkServiceType of the TCP socket, so that these can be customised for each MQTT client.